发布文章开发中。。。

+++ src/main/java/com/example/bbs/entity/Post.java
@@ -18,7 +18,7 @@ public class Post implements Serializable {
 
     private String postContent;
 
-    private Integer postTime;
+    private Long postTime;
 
     private Integer postView;
